requirements.txt
requests
pyyaml
termcolor
urllib3
pip
certifi
idna
setuptools
PySide2